# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'CI4 Board API'
  description: "CodeIgniter 4 기반 헤드리스 게시판 REST API\n\n## 인증\n로그인 후 발급된 `access_token`을 `Authorization: Bearer {token}` 헤더로 전달합니다.\n\n## 기본 응답 형식\n```json\n{ \"success\": true,  \"data\": {},   \"message\": \"\" }\n{ \"success\": false, \"data\": null, \"message\": \"에러 메시지\" }\n```\n\n## API 구조\n| 경로 접두사 | 설명 | 인증 |\n|------------|------|------|\n| `/api/v1/*` | 사용자·게시판·쪽지·CMS | JWT (일부 공개) |\n| `/api/admin/v1/*` | 관리자 전용 | Admin JWT (role: superadmin/manager) |"
  version: 2.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: '개발 서버'
paths:
...
  /api/admin/v1/cms/pages:
    get:
      tags:
        - AdminCMS
      summary: '페이지 목록'
      operationId: 0feaf6f8c6e9d0e9c53a2cba849a48e7
      parameters:
        -
          name: keyword
          in: query
          schema:
            type: string
        -
          name: status
          in: query
          schema:
            type: integer
            enum:
              - 0
              - 1
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: '페이지 목록'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
    post:
      tags:
        - AdminCMS
      summary: '페이지 생성'
      operationId: 5e272438ca287ffdcec4b52abaf10d6a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - slug
                - title
                - contents
              properties:
                slug:
                  description: '영문 소문자·숫자·하이픈'
                  type: string
                  example: about-us
                title:
                  type: string
                contents:
                  type: string
                status:
                  type: integer
                  default: 0
                  enum: [0, 1]
              type: object
      responses:
        '201':
          description: '생성 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
  '/api/admin/v1/cms/pages/{idx}':
    get:
      tags:
        - AdminCMS
      summary: '페이지 단건 조회'
      operationId: 51e80d5a5eddd1e1c0896bf0e6a73896
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: '페이지 상세'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
    put:
      tags:
        - AdminCMS
      summary: '페이지 수정'
      operationId: d6e1bed710cfa5d48443f8cff04b1743
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                slug:
                  type: string
                title:
                  type: string
                contents:
                  type: string
                status:
                  type: integer
                  enum: [0, 1]
              type: object
      responses:
        '200':
          description: '수정 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
    delete:
      tags:
        - AdminCMS
      summary: '페이지 삭제'
      operationId: dce3f2f6732a11831ad8186ec467ee54
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: '삭제 완료'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
...
components:
  schemas:
    ApiResponse:
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
          nullable: true
      type: object
...
  responses:
...
    Unauthorized:
      description: '인증이 필요합니다.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      description: 'JWT Bearer token'
      bearerFormat: JWT
      scheme: bearer
...